delete channel category

+++ src/main/java/net/ruixinglong/suntv/controller/ChannelCategoryController.java
@@ -2,11 +2,15 @@ package net.ruixinglong.suntv.controller;
 
 import com.github.pagehelper.PageInfo;
 import net.ruixinglong.suntv.bean.ChannelCategoryUpdateStatusBean;
+import net.ruixinglong.suntv.bean.PaginationBean;
 import net.ruixinglong.suntv.entity.ChannelCategoryEntity;
+import net.ruixinglong.suntv.entity.ChannelEntity;
 import net.ruixinglong.suntv.exception.BadRequestException;
 import net.ruixinglong.suntv.exception.NotFoundException;
 import net.ruixinglong.suntv.service.ChannelCategoryService;
+import net.ruixinglong.suntv.service.ChannelService;
 import net.ruixinglong.suntv.utils.LocaleMessageUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -23,6 +27,9 @@ public class ChannelCategoryController {
 
     @Resource
     private ChannelCategoryService channelCategoryService;
+
+    @Resource
+    private ChannelService channelService;
 
     @GetMapping("/index")
     public PageInfo<ChannelCategoryEntity> index() throws Exception {
@@ -46,6 +53,7 @@ public class ChannelCategoryController {
             throw new BadRequestException(LocaleMessageUtils.getMsg(bindingResult.getFieldError().getDefaultMessage()));
         }
         request.setCreate_user_id((Integer) httpServletRequest.getAttribute("user_id"));
+        request.setFamily_id((Integer) httpServletRequest.getAttribute("family_id"));
         Integer id = channelCategoryService.create(request);
         ChannelCategoryEntity channelCategoryEntity = channelCategoryService.findOne(id);
         return channelCategoryEntity;
@@ -71,7 +79,11 @@ public class ChannelCategoryController {
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/delete/{id}")
-    public Integer delete(@PathVariable int id) {
+    public Integer delete(@PathVariable int id) throws NotFoundException {
+        List<ChannelEntity> channelList = channelService.findAllByChannelCategoryId(id);
+        if (channelList.size() != 0) {
+            throw new NotFoundException(LocaleMessageUtils.getMsg("channel_category.delete.has_channel"));
+        }
         Integer rows = channelCategoryService.delete(id);
         return rows;
     }
@@ -86,6 +98,13 @@ public class ChannelCategoryController {
     public List<ChannelCategoryEntity> familyAll(HttpServletRequest httpServletRequest) throws Exception {
         int familyId = (Integer) httpServletRequest.getAttribute("family_id");
         List<ChannelCategoryEntity> list = channelCategoryService.findAllByFamilyId(familyId);
+        return list;
+    }
+
+    @GetMapping("/family-index")
+    public PageInfo<ChannelCategoryEntity> familyIndex(Integer pageNum, Integer pageSize,  HttpServletRequest httpServletRequest) throws Exception {
+        int familyId = (Integer) httpServletRequest.getAttribute("family_id");
+        PageInfo<ChannelCategoryEntity> list = channelCategoryService.findAllByFamilyId(familyId, pageNum, pageSize);
         return list;
     }
 }
